minor bug fix

+++ main.cpp
@@ -309,7 +309,7 @@ int main() {
 
 				// calculate percentage of players who have victory
 				double victory_achieved_percentage = (double) players_with_victory.size() 
-					/ players_who_play.size();
+					/ (double) players_who_play.size();
 
 				// initial output
 				cout << "Game: " << game.name() << endl
